added principal for security

+++ Backend/src/main/java/com/Ron/tradingApps/controller/OrderController.java
@@ -5,12 +5,15 @@ import com.Ron.tradingApps.dto.response.OrderResponseDTO;
 import com.Ron.tradingApps.service.BalanceCheckService;
 import com.Ron.tradingApps.service.OrderService;
 import com.Ron.tradingApps.service.PriceCheckService;
+import com.Ron.tradingApps.service.UsernameCheckerService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 @RestController
@@ -24,10 +27,21 @@ public class OrderController {
     private BalanceCheckService balanceCheckService;
     @Autowired
     private PriceCheckService priceCheckService;
+    @Autowired
+    private UsernameCheckerService usernameCheckerService;
 
 
     @PostMapping
-    public ResponseEntity<OrderResponseDTO> createOrder(@RequestBody OrderRequestDTO orderRequestDTO) {
+    public ResponseEntity<OrderResponseDTO> createOrder(@RequestBody OrderRequestDTO orderRequestDTO, Principal principal) {
+        JwtAuthenticationToken token = (JwtAuthenticationToken) principal;
+        String username = token.getTokenAttributes().get("name").toString();
+
+        boolean isUsernameValid = usernameCheckerService.checkUsername(username, orderRequestDTO);
+
+        if (!isUsernameValid) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+
         try {
             OrderResponseDTO orderResponse = orderService.createOrder(orderRequestDTO);
             priceCheckService.checkAndExecuteTransaction(orderResponse);
@@ -41,8 +55,18 @@ public class OrderController {
     }
 
 
+
     @DeleteMapping("/{orderId}")
-    public ResponseEntity<Void> deleteOrder(@PathVariable int orderId) {
+    public ResponseEntity<Void> deleteOrder(@PathVariable int orderId, Principal principal) {
+        JwtAuthenticationToken token = (JwtAuthenticationToken) principal;
+        String username = token.getTokenAttributes().get("name").toString();
+
+        boolean isUsernameValid = usernameCheckerService.checkOrderId(username, orderId);
+
+        if (!isUsernameValid) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+
         try {
             log.info("Attempting to delete order with ID: {}", orderId);
             orderService.deleteOrder(orderId);
@@ -58,7 +82,16 @@ public class OrderController {
     }
 
     @GetMapping("/trader/{userId}")
-    public ResponseEntity<List<OrderResponseDTO>> getAllOrdersForTrader(@PathVariable String userId) {
+    public ResponseEntity<List<OrderResponseDTO>> getAllOrdersForTrader(@PathVariable String userId, Principal principal) {
+        JwtAuthenticationToken token = (JwtAuthenticationToken) principal;
+        String username = token.getTokenAttributes().get("name").toString();
+
+        boolean isUsernameValid = usernameCheckerService.checkUid(username, userId);
+
+        if (!isUsernameValid) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+
         try {
             List<OrderResponseDTO> orders = orderService.findAllOrdersByTrader(userId);
             return new ResponseEntity<>(orders, HttpStatus.OK);
@@ -68,7 +101,11 @@ public class OrderController {
     }
 
     @GetMapping("/trader/{userId}/pending")
-    public ResponseEntity<List<OrderResponseDTO>> getPendingOrdersForTrader(@PathVariable String userId) {
+    public ResponseEntity<List<OrderResponseDTO>> getPendingOrdersForTrader(@PathVariable String userId, Principal principal) {
+        JwtAuthenticationToken token = (JwtAuthenticationToken) principal;
+        String username = token.getTokenAttributes().get("name").toString();
+
+        boolean isUsernameValid = usernameCheckerService.checkUid(username, userId);
         try {
             List<OrderResponseDTO> pendingOrders = orderService.findPendingOrdersByTrader(userId);
             return new ResponseEntity<>(pendingOrders, HttpStatus.OK);
